cmake: Allow Polly to be linked to tools
Fixes the bug CMake Error: install(EXPORT LLVMExports ...) includes target opt which requires target Polly that is not in the export set.
CMake Error: install(EXPORT LLVMExports ...) includes target bugpoint which requires target Polly that is not in the export set.
when building with LINK_POLLY_INTO_TOOLS:Bool=ON

Signed-off-by: Joe Maples <frap129@gmail.com>